requirements.txt
openai
flask
flask-cors
python-dotenv
anthropic
requests
google-genai
streamlit